eval/requirements.txt
# --- AI & LangChain Ecosystem ---
langchain>=0.1.16
langchain-openai>=0.1.3
langchain-community>=0.0.34
langgraph>=0.0.38
openai>=1.20.0

# --- Vector Database ---
chromadb>=0.5.0
rank-bm25==0.2.2

# --- Backend API (FastAPI) ---
fastapi>=0.110.1
uvicorn[standard]>=0.29.0
pydantic>=2.7.0

# --- Frontend ---
streamlit>=1.33.0

# --- Data Processing & Utilities ---
pandas>=2.2.2
requests>=2.31.0
python-dotenv>=1.0.1
emoji==2.15.0
openpyxl==3.1.5
pypdf==6.9.1
pymongo[srv]==3.12
matplotlib==3.10.7

deepeval>=0.21.0
ragas>=0.1.0
datasets>=2.18.0